Image Steganography using PILL module from python
Image Steganography is the art of hiding a file in another file. These files can be image, audio and video. This is mainly practiced to hide the fact that communication is taking place or has taken place. To be more precise, this technology leaves no trace or suspicion . This will be helpful in cases where you don’t want to send information openly or even in an encrypted form. Cryptography and Steganography are very much kind of siblings. While cryptography provides privacy, steganography provides privacy as well as secrecy. Different applications have different requirements of the steganography techniques to be used. For example, some applications may require absolute invisibility of the secret information, while other requires a larger secret message to be hidden. This paper intends to give you an overview about the concept of Image Steganography, it uses and techniques to implement it.